Move Errors component into its own components sub-directory
