+++ mathesar_ui/src/components/errors/Errors.svelte
