+++ mathesar_ui/src/systems/permissions/PermissionsModalContent.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { _ } from 'svelte-i18n';
 
-  import Errors from '@mathesar/components/Errors.svelte';
+  import Errors from '@mathesar/components/errors/Errors.svelte';
   import {
     ImmutableMap,
     Spinner,
